+++ apple-codesign/src/yubikey.rs
@@ -6,7 +6,7 @@
 
 use {
     crate::AppleCodesignError,
-    log::warn,
+    log::{error, warn},
     std::ops::DerefMut,
     std::sync::{Arc, Mutex, MutexGuard},
     x509_certificate::{
@@ -14,9 +14,9 @@ use {
         X509CertificateError,
     },
     yubikey::{
-        certificate::Certificate as YkCertificate,
-        piv::{AlgorithmId, SlotId},
-        Error as YkError, YubiKey as RawYubiKey,
+        certificate::{CertInfo, Certificate as YkCertificate},
+        piv::{import_ecc_key, import_rsa_key, AlgorithmId, SlotId},
+        Error as YkError, MgmKey, YubiKey as RawYubiKey, {PinPolicy, TouchPolicy},
     },
     zeroize::Zeroizing,
 };
@@ -24,9 +24,61 @@ use {
 /// A function that will attempt to resolve the PIN to unlock a YubiKey.
 pub type PinCallback = fn() -> Result<Vec<u8>, AppleCodesignError>;
 
+fn algorithm_from_certificate(
+    cert: &CapturedX509Certificate,
+) -> Result<AlgorithmId, X509CertificateError> {
+    let key_algorithm = cert
+        .key_algorithm()
+        .ok_or(X509CertificateError::UnknownKeyAlgorithm(format!(
+            "{:?}",
+            cert.key_algorithm_oid()
+        )))?;
+
+    match key_algorithm {
+        KeyAlgorithm::Rsa => match cert.rsa_public_key_data()?.modulus.as_slice().len() {
+            129 => Ok(AlgorithmId::Rsa1024),
+            257 => Ok(AlgorithmId::Rsa2048),
+            _ => Err(X509CertificateError::Other(
+                "unable to determine RSA key algorithm".into(),
+            )),
+        },
+        KeyAlgorithm::Ed25519 => Err(X509CertificateError::UnknownKeyAlgorithm(
+            "unable to use ed25519 keys with smartcards".into(),
+        )),
+        KeyAlgorithm::Ecdsa(curve) => match curve {
+            EcdsaCurve::Secp256r1 => Ok(AlgorithmId::EccP256),
+            EcdsaCurve::Secp384r1 => Ok(AlgorithmId::EccP384),
+        },
+    }
+}
+
+/// Describes the needed authentication for an operation.
+pub enum RequiredAuthentication {
+    Pin,
+    ManagementKey,
+    ManagementKeyAndPin,
+}
+
+impl RequiredAuthentication {
+    pub fn requires_pin(&self) -> bool {
+        match self {
+            Self::Pin | Self::ManagementKeyAndPin => true,
+            Self::ManagementKey => false,
+        }
+    }
+
+    pub fn requires_management_key(&self) -> bool {
+        match self {
+            Self::ManagementKey | Self::ManagementKeyAndPin => true,
+            Self::Pin => false,
+        }
+    }
+}
+
 fn attempt_authenticated_operation<T>(
     yk: &mut RawYubiKey,
     op: impl Fn(&mut RawYubiKey) -> Result<T, AppleCodesignError>,
+    required_authentication: RequiredAuthentication,
     get_device_pin: Option<&PinCallback>,
 ) -> Result<T, AppleCodesignError> {
     const MAX_ATTEMPTS: u8 = 3;
@@ -44,25 +96,44 @@ fn attempt_authenticated_operation<T>(
                     return Err(AppleCodesignError::SmartcardFailedAuthentication);
                 }
 
-                warn!("device refused to sign due to authentication error");
+                warn!("device refused operation due to authentication error");
 
-                if let Some(pin_cb) = get_device_pin {
-                    let pin = Zeroizing::new(pin_cb().map_err(|e| {
-                        X509CertificateError::Other(format!("error retrieving device pin: {}", e))
-                    })?);
-
-                    match yk.verify_pin(&pin) {
+                if required_authentication.requires_management_key() {
+                    match yk.authenticate(MgmKey::default()) {
                         Ok(()) => {
-                            warn!("pin verification successful");
+                            warn!("management key authentication successful");
                         }
                         Err(e) => {
-                            warn!("pin verification failure: {}", e);
+                            error!("management key authentication failure: {}", e);
                             continue;
                         }
                     }
-                } else {
-                    warn!("unable to retrieve device pin; future attempts will fail; giving up");
-                    return Err(AppleCodesignError::SmartcardFailedAuthentication);
+                }
+
+                if required_authentication.requires_pin() {
+                    if let Some(pin_cb) = get_device_pin {
+                        let pin = Zeroizing::new(pin_cb().map_err(|e| {
+                            X509CertificateError::Other(format!(
+                                "error retrieving device pin: {}",
+                                e
+                            ))
+                        })?);
+
+                        match yk.verify_pin(&pin) {
+                            Ok(()) => {
+                                warn!("pin verification successful");
+                            }
+                            Err(e) => {
+                                error!("pin verification failure: {}", e);
+                                continue;
+                            }
+                        }
+                    } else {
+                        warn!(
+                            "unable to retrieve device pin; future attempts will fail; giving up"
+                        );
+                        return Err(AppleCodesignError::SmartcardFailedAuthentication);
+                    }
                 }
             }
             Err(e) => {
@@ -156,6 +227,150 @@ impl YubiKey {
                 }
             }))
     }
+
+    fn import_rsa_key(
+        &mut self,
+        p: &[u8],
+        q: &[u8],
+        cert: &CapturedX509Certificate,
+        slot: SlotId,
+        touch_policy: TouchPolicy,
+        pin_policy: PinPolicy,
+    ) -> Result<(), AppleCodesignError> {
+        let slot_pretty = hex::encode([u8::from(slot)]);
+
+        let public_key_data = cert.rsa_public_key_data()?;
+
+        let algorithm = match public_key_data.modulus.as_slice().len() {
+            129 => AlgorithmId::Rsa1024,
+            257 => AlgorithmId::Rsa2048,
+            _ => {
+                return Err(X509CertificateError::Other(
+                    "unable to determine RSA key algorithm".into(),
+                )
+                .into());
+            }
+        };
+
+        warn!(
+            "attempting import of {:?} private key to slot {}",
+            algorithm, slot_pretty
+        );
+
+        let mut yk = self.inner()?;
+
+        attempt_authenticated_operation(
+            yk.deref_mut(),
+            |yk| {
+                let rsa_key = ::yubikey::piv::RsaKeyData::new(&p, &q);
+
+                import_rsa_key(yk, slot, algorithm, rsa_key, touch_policy, pin_policy)?;
+
+                Ok(())
+            },
+            RequiredAuthentication::ManagementKeyAndPin,
+            self.pin_callback.as_ref(),
+        )?;
+
+        Ok(())
+    }
+
+    fn import_ecdsa_key(
+        &mut self,
+        private_key: &[u8],
+        cert: &CapturedX509Certificate,
+        slot: SlotId,
+        touch_policy: TouchPolicy,
+        pin_policy: PinPolicy,
+    ) -> Result<(), AppleCodesignError> {
+        let slot_pretty = hex::encode([u8::from(slot)]);
+
+        let algorithm = algorithm_from_certificate(cert)?;
+
+        warn!(
+            "attempting import of ECDSA private key to slot {}",
+            slot_pretty
+        );
+
+        let mut yk = self.inner()?;
+
+        attempt_authenticated_operation(
+            yk.deref_mut(),
+            |yk| {
+                import_ecc_key(yk, slot, algorithm, private_key, touch_policy, pin_policy)?;
+
+                Ok(())
+            },
+            RequiredAuthentication::ManagementKeyAndPin,
+            self.pin_callback.as_ref(),
+        )?;
+
+        Ok(())
+    }
+
+    /// Attempt to import a private key and certificate into the YubiKey.
+    pub fn import_key(
+        &mut self,
+        slot: SlotId,
+        key: &dyn Sign,
+        cert: &CapturedX509Certificate,
+        touch_policy: TouchPolicy,
+        pin_policy: PinPolicy,
+    ) -> Result<(), AppleCodesignError> {
+        let slot_pretty = hex::encode([u8::from(slot)]);
+
+        let certificate = YkCertificate::from_bytes(cert.encode_der()?)?;
+
+        match cert.key_algorithm() {
+            Some(KeyAlgorithm::Rsa) => {
+                let (p, q) = key.rsa_primes()?.ok_or_else(|| {
+                    X509CertificateError::Other(
+                        "could not locate RSA private key parameters".into(),
+                    )
+                })?;
+
+                self.import_rsa_key(&p, &q, cert, slot, touch_policy, pin_policy)?;
+            }
+            Some(KeyAlgorithm::Ecdsa(_)) => {
+                let private_key = key.private_key_data().ok_or_else(|| {
+                    X509CertificateError::Other("could not retrieve private key data".into())
+                })?;
+
+                self.import_ecdsa_key(&private_key, cert, slot, touch_policy, pin_policy)?;
+            }
+            Some(algorithm) => {
+                return Err(AppleCodesignError::CertificateUnsupportedKeyAlgorithm(
+                    algorithm,
+                ));
+            }
+            None => {
+                return Err(X509CertificateError::UnknownKeyAlgorithm("unknown".into()).into());
+            }
+        }
+
+        warn!(
+            "successfully wrote private key to slot {}; proceeding to write certificate",
+            slot_pretty
+        );
+
+        let mut yk = self.inner()?;
+
+        // The key is imported! Now try to write the public certificate next to it.
+        attempt_authenticated_operation(
+            yk.deref_mut(),
+            |yk| {
+                certificate.write(yk, slot, CertInfo::Uncompressed)?;
+
+                Ok(())
+            },
+            RequiredAuthentication::ManagementKeyAndPin,
+            self.pin_callback.as_ref(),
+        )?;
+
+        warn!("successfully wrote certificate to slot {}", slot_pretty);
+
+        Ok(())
+    }
 }
 
 /// Entity for creating signatures using a certificate in a given PIV slot.
@@ -170,34 +385,7 @@ pub struct CertificateSigner {
 
 impl Sign for CertificateSigner {
     fn sign(&self, message: &[u8]) -> Result<(Vec<u8>, SignatureAlgorithm), X509CertificateError> {
-        let key_algorithm =
-            self.cert
-                .key_algorithm()
-                .ok_or(X509CertificateError::UnknownKeyAlgorithm(format!(
-                    "{:?}",
-                    self.cert.key_algorithm_oid()
-                )))?;
-
-        let algorithm_id = match key_algorithm {
-            KeyAlgorithm::Rsa => match self.cert.rsa_public_key_data()?.modulus.as_slice().len() {
-                129 => AlgorithmId::Rsa1024,
-                257 => AlgorithmId::Rsa2048,
-                _ => {
-                    return Err(X509CertificateError::Other(
-                        "unable to determine RSA key algorithm".into(),
-                    ));
-                }
-            },
-            KeyAlgorithm::Ed25519 => {
-                return Err(X509CertificateError::UnknownKeyAlgorithm(
-                    "unable to use ed25519 keys with smartcards".into(),
-                ));
-            }
-            KeyAlgorithm::Ecdsa(curve) => match curve {
-                EcdsaCurve::Secp256r1 => AlgorithmId::EccP256,
-                EcdsaCurve::Secp384r1 => AlgorithmId::EccP384,
-            },
-        };
+        let algorithm_id = algorithm_from_certificate(&self.cert)?;
 
         let signature_algorithm =
             self.cert
@@ -239,6 +427,7 @@ impl Sign for CertificateSigner {
 
                 Ok((signature.to_vec(), signature_algorithm))
             },
+            RequiredAuthentication::Pin,
             self.pin_callback.as_ref(),
         )
         .map_err(|e| X509CertificateError::Other(format!("code sign error: {:?}", e)))
@@ -251,6 +440,15 @@ impl Sign for CertificateSigner {
                 self.cert.signature_algorithm_oid()
             )),
         )?)
+    }
+
+    fn private_key_data(&self) -> Option<&[u8]> {
+        // We never have access to private keys stored on hardware devices.
+        None
+    }
+
+    fn rsa_primes(&self) -> Result<Option<(Vec<u8>, Vec<u8>)>, X509CertificateError> {
+        Ok(None)
     }
 }